backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.4
pydantic-settings>=2.3.4
firebase-admin>=6.5.0
google-generativeai>=0.7.0
chromadb>=0.5.0
google-cloud-storage>=2.17.0
celery[redis]>=5.4.0
redis>=5.0.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
aiosqlite>=0.20.0
lightgbm>=4.3.0
joblib>=1.4.0
pandas>=2.2.0
scikit-learn>=1.5.0
numpy>=1.26.0
alembic>=1.13.0
pytest>=8.2.2
httpx>=0.27.0
beautifulsoup4>=4.12.3
pytest-asyncio>=0.23.7
